08/04/2015
Addition of InternalPane.
Tackling DatabaseConnectionDialog class to insert data into the new
InternalPane will need to revert DatabaseConnectionDialog if problems
begin to occur.
OperationsTest will soon be discontinued from the use of the program and
all information that was once written into the OperationsPane will need
to be moved over to the new InternalPane.
Also addition of new documentation explaining the advantages and
disadvantages of the internal pane and any implementation issues that
may arise, This must be added to in the future as the internalPane
begins to grow